Board resize, lock cards for editing, bugfixes, download to csv
fixed clients joining, leaving, renaming

Npm install will install all packages for API and UI

Adds `scrumblr` init.d script and ensures that UI packages are installed  (#10)

Added react-csv and download to csv button (#11)

Feature - locked cards for editing, display which client is editing a card

Interesting note about react-csv. There's a known "issue" with trying to download data that is loaded asynchronously. CSVLink component loads with a static set of data. Scrumblr loads each cards as it receives them over the socket, so as soon as one card is present in the store, the CSVLink is mounted with the first card.

To mitigate this, I make the user click the 'Download CSV' div, which is bound to boolean property 'download' in the state. The link 'Export CSV' will appear when the download property is set to true. Once the user downloads the data, I set the download property back to false, so that they must regenerate the link again, this will ensure that the link stays up to date with the data on the board.

try not to obstruct close button

Rearrange cards when board resizes

remove serviceworker

upgraded react-scripts